+++ src/main/java/br/com/cattle_control/starter/repository/PeopleRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.cattle_control.starter.model.People;
 
-public interface PeopleRepository extends JpaRepository<People, Long> {
+public interface PeopleRepository extends JpaRepository<People, Integer> {
     
 }
